requirements.txt
Django>=4,<5
pillow==11.0.0
python-dotenv==1.0.1
python-telegram-bot==13.7
